Разбил общую модель на несколько отдельных файлов
